+++ src/main/java/cn/elytra/mod/nomi_horizons/NomiHorizons.java
@@ -4,16 +4,21 @@ import crazypants.enderio.machines.init.MachineObject;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.item.ItemBlock;
+import net.minecraftforge.common.config.Config;
+import net.minecraftforge.common.config.ConfigManager;
+import net.minecraftforge.fml.client.event.ConfigChangedEvent;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import xyz.vsngamer.elevator.init.Registry;
 
 @Mod(modid = Tags.MOD_ID, name = Tags.MOD_NAME, version = Tags.VERSION, dependencies = NomiHorizons.DEPS)
+@Mod.EventBusSubscriber(modid = Tags.MOD_ID)
 public class NomiHorizons {
 
     public static final Logger LOG = LogManager.getLogger(Tags.MOD_NAME);
@@ -33,6 +38,13 @@ public class NomiHorizons {
     @EventHandler
     public void postInit(FMLPostInitializationEvent event) {
         loadCrossModCompat();
+    }
+
+    @SubscribeEvent
+    public static void onConfigSync(ConfigChangedEvent.OnConfigChangedEvent event) {
+        if(event.getModID().equals(Tags.MOD_ID)) {
+            ConfigManager.sync(Tags.MOD_ID, Config.Type.INSTANCE);
+        }
     }
 
     private void loadCrossModCompat() {