+++ src/main/java/cn/elytra/mod/nomi_horizons/NomiHorizonsAPI.java
@@ -1,9 +1,14 @@
 package cn.elytra.mod.nomi_horizons;
 
+import cn.elytra.mod.nomi_horizons.config.NomiHorizonsConfig;
+import codechicken.lib.util.ArrayUtils;
 import com.google.common.collect.Lists;
+import gregtech.api.metatileentity.multiblock.CleanroomType;
 import gregtech.api.metatileentity.multiblock.MultiblockControllerBase;
 import gregtech.api.pattern.TraceabilityPredicate;
 import net.minecraft.block.state.IBlockState;
+import net.minecraftforge.common.config.Config;
+import net.minecraftforge.common.config.ConfigManager;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
@@ -60,6 +65,16 @@ public class NomiHorizonsAPI {
     @ApiStatus.Internal
     public int getMaxElevatorCount() {
         return maxElevatorCount;
+    }
+
+    @ApiStatus.Internal
+    public boolean isCleanroomPardoned() {
+        return NomiHorizonsConfig.pardonCleanroom;
+    }
+
+    @ApiStatus.Internal
+    public boolean isSterileCleanroomPardoned() {
+        return NomiHorizonsConfig.pardonSterileCleanroom;
     }
 
 }