+++ src/main/java/cn/elytra/mod/nomi_horizons/config/NomiHorizonsConfig.java
@@ -1,0 +1,17 @@
+package cn.elytra.mod.nomi_horizons.config;
+
+import cn.elytra.mod.nomi_horizons.Tags;
+import net.minecraftforge.common.config.Config;
+
+@Config(modid = Tags.MOD_ID)
+public class NomiHorizonsConfig {
+
+    @Config.LangKey("nomi_horizons.config.pardon_cleanroom")
+    @Config.Comment("True to bypass the Cleanroom check for multiblocks.")
+    public static boolean pardonCleanroom = true;
+
+    @Config.LangKey("nomi_horizons.config.pardon_sterile_cleanroom")
+    @Config.Comment("True to bypass the Sterile Cleanroom check for multiblocks.")
+    public static boolean pardonSterileCleanroom = false;
+
+}